requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docutils==0.20.1
idna==3.4
Kivy==2.2.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
numpy==1.24.3
pandas==2.0.1
Pygments==2.15.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
